Fix bug in deleting queries.

+++ __init__.py
@@ -535,7 +535,6 @@ def edit_category(category_id):
 def delete_category(category_id):
     """Delete a category."""
     category = session.query(Category).filter_by(id=category_id).first()
-    items = session.query(Item).filter_by(category_id=category.id).all()
 
     if 'username' not in login_session:
         flash("Please log in to continue.")
@@ -552,12 +551,16 @@ def delete_category(category_id):
         return redirect(url_for('home'))
 
     if request.method == 'POST':
-        session.delete(items)
-        session.commit()
-        session.delete(category)
-        session.commit()
-        flash("Category successfully deleted!")
-        return redirect(url_for('home'))
+        total = session.query(Item).filter_by(category_id=category.id).count()
+        # Delete only if the category has no items.
+        if total == 0:
+            session.delete(category)
+            session.commit()
+            flash("Category successfully deleted!")
+            return redirect(url_for('home'))
+        else:
+            flash("We are unable to process your request right now.")
+            return redirect(url_for('home'))
     else:
         return render_template("delete_category.html", category=category)
